shuffle some more stuff around between modules
