+++ persistent/persistent.cabal
@@ -81,6 +81,7 @@ library
     Database.Persist.PersistValue
     Database.Persist.Quasi
     Database.Persist.Quasi.PersistSettings
+    Database.Persist.Quasi.PersistSettings.Internal
     Database.Persist.Quasi.Internal
     Database.Persist.Quasi.Internal.ModelParser
     Database.Persist.Sql
